common: Move AuthTag from common to the library
Move the AuthTag structure from the common.rs file to the Keylime
library in crypto::auth_tag
